+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Adds four circle blueprint
 - Adds brush size control
 - Adds modulo attribute to circles
+- Adds loading by index
+- Adds modulo controls
 
 ## [1.0.0]
 ### Added